refactor:#54 query params 상수 추가
